+++ rstar/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+- Add `RTree::drain_*` methods to remove and drain selected items. ([PR](https://github.com/georust/rstar/pull/77))
+
+## Changed
+- Expose all iterator types in `crate::iterators` module ([PR](https://github.com/georust/rstar/pull/77))
+
 # 0.9.1
 
 ## Added